requirements.txt
# local package
-e .

# external requirements
click~=8.1.3
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

tensorflow~=2.9.2
pandas~=1.4.2
setuptools~=63.2.0
pydantic~=1.10.2